requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
groq==0.4.2
python-dotenv==1.0.0
Werkzeug==2.3.7